refactor profile's facebook helper methods
